reduce cloudflare functions build size
